22.5.0.0

Improvement:

Created Revit.IFC.Import.Core DLL to house Navis processor and interface.
Added support for Sound power units on IFC export.
Corrected the IFC GUID parameter value on export if the value used was different than the value set for the element.
Improved Door and Window property units for Navisworks.
Created DLL for UI to select IFC Entity to be used for Entity selection.
Disallow openings and connections for Walls that are not exported as IfcElements.
Implemented removing of isolated points in polygonal faceset.
Implemented IfcDistributionCircuit export.
Updated ability of Navis to create ProjectInformation before adding parameters.
Improved face margin.
Improved linking of extrusions with slivery outlines.
Added support for linking IFC files with the IfcBlock entity.
Added non-optional attributes export for generic type.

UI bugfix:

Removed truncated strings from some localized IFC dialogs.
Fixed error during renaming of IFC export setup error.
Improved localization of the current selected setup dropdown in the IFC export dialog box.
Improved localization of some IFC export dialog options.
Fixing GLOB issue in some labels and buttons

Export bugfix:

Improved export of some walls to IFC whose end is cut by an in-place void.
Improved export of curved railings material.
Corrected units for some properties of rooms and spaces in schedules exported to IFC as property sets.
Fixed the center line location of beams with trimmed ends exported to IFC 2x3.
Fix for Column family geometry export.
Improved export of some clipped structural beams to IFC.
Improved export of walls containing adjacent openings to IFC.
Improved the export of walls with extensions cut by wall sweeps to IFC.
Fixed export of roofs as polygonal face sets.
Added opening creation in floor.
Improved the creation of openings for some walls export to IFC with the option to split walls by level selected.
Fixed issue with unintended unit conversion for IFC property values.
Fixed export of thermal transmittance parameters to IFC.
Fixed CrossSectionArea property.
Fixed material display of roofs.
Fixed pressure properties scaling.
Fix for shifted openings on floor.
Fixed curtain walls geometry export

Import bugfix:

Included wall in linking of IFC walls when they contained excluded doors and windows.
Fixed linking of axes with invalid data.
Improved Openings cuts processing when linking IFC.
Fixed IfcIndexedPolyCurve linking.
Fixed linking of some triangulated face sets.
Improved support for linking some invalid IFC files
Improved linking of IFC files that contained geometry with degenerate faces.

+++ Source/Revit.IFC.Import/Processors/IFCDefaultProcessor.cs
@@ -1,18 +1,19 @@
 ﻿using System.Collections.Generic;
 
 using Autodesk.Revit.DB;
+using Revit.IFC.Import.Core;
 
 namespace Revit.IFC.Import
 {
    public class IFCDefaultProcessor : IIFCFileProcessor
    {
-      public bool CreateOrUpdateElement(int ifcId, string ifcGuid, string ifcEntityType, int categoryId,
+      public bool CreateOrUpdateElement(int ifcId, string ifcGuid, string ifcEntityType, long categoryId,
          IList<GeometryObject> geomObjs)
       {
          return false;
       }
 
-      public void CreateElementType(int ifcId, string ifcGuid, string ifcEntityType, int categoryId)
+      public void CreateElementType(int ifcId, string ifcGuid, string ifcEntityType, long categoryId)
       {
       }
